Time: 9 ms (6.27%), Space: 48.9 MB (49.45%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.java
@@ -14,7 +14,7 @@ class Solution {
         Arrays.sort(merged);
         int total = merged.length;
         if(total %2 == 1){
-            return (double) merged[total/2];
+            return merged[total/2];
         }else{
             int mid1 = merged[total/2 - 1];
             int mid2 = merged[total/2];
